Don't fail GS1 message parsing on inner format error (#169)
This is a breaking change, as we no longer raise a `ParseError` when inner
formats (like GLN, GTIN, and SSCC) inside a GS1 element string fails parsing or
validation. Instead we set the associated error fields on the returned
`GS1ElementString`, e.g. `gln_error`, `gtin_error`, or `sscc_error`.

The reason for this change is that there might be use cases where you're
interested in the other parts of the GS1 message than the one that fails
validation.